Conflicts:
	BeatYourPace/bin/BeatYourPace.apk
	BeatYourPace/bin/classes.dex
	BeatYourPace/bin/resources.ap_

+++ BeatYourPace/src/com/GC01/BeatYourPace/Database/DataModel.java
@@ -100,7 +100,7 @@ public class DataModel {
 		return initialPrefPace;
 	}
 
-	public void setInitialPrefPace(float initialPrefPace) {
+	public void setInitialPrefPace(int initialPrefPace) {
 		this.initialPrefPace = initialPrefPace;
 	}
